feat: 게임중 예외 발생시 명령어 재 입력 기능
