+++ src/main/java/chess/domain/piece/King.java
@@ -1,24 +1,23 @@
-package chess.piece;
+package chess.domain.piece;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import chess.position.Position;
+import chess.domain.position.Position;
 
-public class Knight extends Piece {
-	private static final String INITIAL_CHARACTER = "N";
+public class King extends Piece {
+	private static final String INITIAL_CHARACTER = "K";
 
-	public Knight(Team team) {
+	public King(Team team) {
 		super(team);
 	}
 
 	@Override
 	public List<Position> findMoveModeTrace(Position from, Position to) {
-		if (from.isNotMultiplicationOfDifferenceBetweenFileAndRankIsTwo(to)) {
+		if (from.isNotDistanceOneSquare(to)) {
 			throw new IllegalArgumentException("해당 위치로 이동할 수 없습니다.");
 		}
-		return new ArrayList<>(Collections.emptyList());
+		return Collections.emptyList();
 	}
 
 	@Override
@@ -27,7 +26,12 @@ public class Knight extends Piece {
 	}
 
 	@Override
+	public boolean isKing() {
+		return true;
+	}
+
+	@Override
 	public double getScore() {
-		return 2.5;
+		return 0;
 	}
 }
