+++ src/main/java/chess/domain/piece/Knight.java
@@ -1,23 +1,24 @@
-package chess.piece;
+package chess.domain.piece;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import chess.position.Position;
+import chess.domain.position.Position;
 
-public class King extends Piece {
-	private static final String INITIAL_CHARACTER = "K";
+public class Knight extends Piece {
+	private static final String INITIAL_CHARACTER = "N";
 
-	public King(Team team) {
+	public Knight(Team team) {
 		super(team);
 	}
 
 	@Override
 	public List<Position> findMoveModeTrace(Position from, Position to) {
-		if (from.isNotDistanceOneSquare(to)) {
+		if (from.isNotMultiplicationOfDifferenceBetweenFileAndRankIsTwo(to)) {
 			throw new IllegalArgumentException("해당 위치로 이동할 수 없습니다.");
 		}
-		return Collections.emptyList();
+		return new ArrayList<>(Collections.emptyList());
 	}
 
 	@Override
@@ -26,12 +27,7 @@ public class King extends Piece {
 	}
 
 	@Override
-	public boolean isKing() {
-		return true;
-	}
-
-	@Override
 	public double getScore() {
-		return 0;
+		return 2.5;
 	}
 }
