+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,22 +1,22 @@
-package chess.piece;
+package chess.domain.piece;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
-import chess.position.File;
-import chess.position.Position;
-import chess.position.Rank;
+import chess.domain.position.File;
+import chess.domain.position.Position;
+import chess.domain.position.Rank;
 
-public class Rook extends Piece {
-	private static final String INITIAL_CHARACTER = "R";
+public class Queen extends Piece {
+	private static final String INITIAL_CHARACTER = "Q";
 
-	public Rook(Team team) {
+	public Queen(Team team) {
 		super(team);
 	}
 
 	@Override
 	public List<Position> findMoveModeTrace(Position from, Position to) {
-		if (from.isNotStraight(to)) {
+		if (from.isNotStraight(to) && from.isNotDiagonal(to)) {
 			throw new IllegalArgumentException("해당 위치로 이동할 수 없습니다.");
 		}
 		if (from.isSameRank(to)) {
@@ -24,6 +24,9 @@ public class Rook extends Piece {
 			return files.stream()
 				.map(file -> Position.of(file, from.getRank()))
 				.collect(Collectors.toList());
+		}
+		if (from.isDiagonal(to)) {
+			return Position.findDiagonalTrace(from, to);
 		}
 		List<Rank> ranks = Rank.valuesBetween(from.getRank(), to.getRank());
 		return ranks.stream()
@@ -38,6 +41,6 @@ public class Rook extends Piece {
 
 	@Override
 	public double getScore() {
-		return 5;
+		return 9;
 	}
 }
