+++ src/main/java/chess/domain/piece/Rook.java
@@ -1,22 +1,22 @@
-package chess.piece;
+package chess.domain.piece;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
-import chess.position.File;
-import chess.position.Position;
-import chess.position.Rank;
+import chess.domain.position.File;
+import chess.domain.position.Position;
+import chess.domain.position.Rank;
 
-public class Queen extends Piece {
-	private static final String INITIAL_CHARACTER = "Q";
+public class Rook extends Piece {
+	private static final String INITIAL_CHARACTER = "R";
 
-	public Queen(Team team) {
+	public Rook(Team team) {
 		super(team);
 	}
 
 	@Override
 	public List<Position> findMoveModeTrace(Position from, Position to) {
-		if (from.isNotStraight(to) && from.isNotDiagonal(to)) {
+		if (from.isNotStraight(to)) {
 			throw new IllegalArgumentException("해당 위치로 이동할 수 없습니다.");
 		}
 		if (from.isSameRank(to)) {
@@ -24,9 +24,6 @@ public class Queen extends Piece {
 			return files.stream()
 				.map(file -> Position.of(file, from.getRank()))
 				.collect(Collectors.toList());
-		}
-		if (from.isDiagonal(to)) {
-			return Position.findDiagonalTrace(from, to);
 		}
 		List<Rank> ranks = Rank.valuesBetween(from.getRank(), to.getRank());
 		return ranks.stream()
@@ -41,6 +38,6 @@ public class Queen extends Piece {
 
 	@Override
 	public double getScore() {
-		return 9;
+		return 5;
 	}
 }
