+++ src/main/java/chess/domain/position/Position.java
@@ -1,7 +1,6 @@
-package chess.position;
+package chess.domain.position;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
