+++ src/main/java/chess/domain/position/Rank.java
@@ -1,4 +1,4 @@
-package chess.position;
+package chess.domain.position;
 
 import java.util.Arrays;
 import java.util.Comparator;
