+++ src/test/java/chess/domain/piece/RookTest.java
@@ -1,8 +1,8 @@
-package chess.piece;
+package chess.domain.piece;
 
-import static chess.piece.Team.*;
-import static chess.position.File.*;
-import static chess.position.Rank.*;
+import static chess.domain.piece.Team.*;
+import static chess.domain.position.File.*;
+import static chess.domain.position.Rank.*;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.Arrays;
@@ -16,44 +16,40 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import chess.position.Position;
+import chess.domain.position.Position;
 
-public class BishopTest {
+public class RookTest {
 	@ParameterizedTest
 	@MethodSource("startDestinationTraceProvider")
-	void bishopPathTest(Position start, Position destination, List<Position> trace) {
-		Bishop bishop = new Bishop(BLACK);
-		List<Position> actual = bishop.findMoveModeTrace(start, destination);
+	void rookPathTest(Position start, Position destination, List<Position> trace) {
+		Rook rook = new Rook(BLACK);
+		List<Position> actual = rook.findMoveModeTrace(start, destination);
 		assertThat(actual).isEqualTo(trace);
 	}
 
 	private static Stream<Arguments> startDestinationTraceProvider() {
 		return Stream.of(
-			Arguments.of(Position.of(C, SIX), Position.of(F, THREE),
-				Arrays.asList(Position.of(D, FIVE), Position.of(E, FOUR))),
-			Arguments.of(Position.of(F, THREE), Position.of(C, SIX),
-				Arrays.asList(Position.of(E, FOUR), Position.of(D, FIVE))),
-			Arguments.of(Position.of(C, THREE), Position.of(F, SIX),
-				Arrays.asList(Position.of(D, FOUR), Position.of(E, FIVE))),
-			Arguments.of(Position.of(F, SIX), Position.of(C, THREE),
-				Arrays.asList(Position.of(E, FIVE), Position.of(D, FOUR)))
+			Arguments.of(Position.of(B, FOUR), Position.of(B, SIX),
+				Arrays.asList(Position.of(B, FIVE))),
+			Arguments.of(Position.of(B, FOUR), Position.of(E, FOUR),
+				Arrays.asList(Position.of(C, FOUR), Position.of(D, FOUR)))
 		);
 	}
 
 	@DisplayName("허용되지 않은 출발위치와 도착위치인 경우 예외가 발생하는지 테스트")
 	@Test
 	void invalidMovementTest() {
-		Bishop bishop = new Bishop(BLACK);
-		assertThatThrownBy(() -> bishop.findMoveModeTrace(Position.of(A, ONE), Position.of(B, ONE)))
+		Rook rook = new Rook(BLACK);
+		assertThatThrownBy(() -> rook.findMoveModeTrace(Position.of(A, ONE), Position.of(B, TWO)))
 			.isInstanceOf(IllegalArgumentException.class)
 			.hasMessage("해당 위치로 이동할 수 없습니다.");
 	}
 
 	@DisplayName("체스말의 팀과 종류에 따라 심볼이 반환된다.")
 	@ParameterizedTest
-	@CsvSource(value = {"BLACK,B", "WHITE,b"})
+	@CsvSource(value = {"BLACK,R", "WHITE,r"})
 	void getSymbolTest(Team team, String expected) {
-		Piece piece = new Bishop(team);
+		Piece piece = new Rook(team);
 		assertThat(piece.getSymbol()).isEqualTo(expected);
 	}
 }
