+++ src/test/java/chess/domain/position/RankTest.java
@@ -1,4 +1,4 @@
-package chess.position;
+package chess.domain.position;
 
 import static org.assertj.core.api.Assertions.*;
